Additional casts to avoid compiler warnings, originally from a MS compiler, but also given by gcc if you turn on enough warnings.
git-svn-id: svn://vcs.exim.org/pcre/code/trunk@904 2f5784b3-3f2a-0410-8824-cb99058d5e15

+++ pcre_compile.c
@@ -3067,7 +3067,7 @@ if (next >= 0) switch(op_code)
     }
   else
 #endif  /* SUPPORT_UTF */
-  return (c != TABLE_GET(next, cd->fcc, next));  /* Non-UTF-8 mode */
+  return (c != TABLE_GET((unsigned int)next, cd->fcc, next));  /* Non-UTF-8 mode */
 
   /* For OP_NOT and OP_NOTI, the data is always a single-byte character. These
   opcodes are not used for multi-byte characters, because they are coded using
@@ -3092,7 +3092,7 @@ if (next >= 0) switch(op_code)
     }
   else
 #endif  /* SUPPORT_UTF */
-  return (c == TABLE_GET(next, cd->fcc, next));  /* Non-UTF-8 mode */
+  return (c == (int)(TABLE_GET((unsigned int)next, cd->fcc, next)));  /* Non-UTF-8 mode */
 
   /* Note that OP_DIGIT etc. are generated only when PCRE_UCP is *not* set.
   When it is set, \d etc. are converted into OP_(NOT_)PROP codes. */
@@ -4571,7 +4571,7 @@ for (;; ptr++)
 #endif
           {
           unsigned int othercase;
-          if ((othercase = UCD_OTHERCASE(c)) != c)
+          if ((int)(othercase = UCD_OTHERCASE(c)) != c)
             {
             *class_uchardata++ = XCL_SINGLE;
             class_uchardata += PRIV(ord2utf)(othercase, class_uchardata);
